requirements.txt
# requirements.txt
aiohttp
python-opensky
geopy
requests
python-dotenv
folium
openmeteo_requests
requests_cache
retry_requests
pandas
